% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hnc_df.R
\docType{data}
\name{hnc_df}
\alias{hnc_df}
\title{Head and Neck cancer Quality of Life data.}
\format{
A data frame with 60 rows and 2 variables:
\describe{
  \item{ID}{Participant's identification}
  \item{time}{Time Variable}
  \item{event}{status as Variable}
  \item{arm}{Therapeutic Arm}
  \item{HN_Q31}{HNC Cancer Quality of Q31 Question}
  \item{HN_Q32}{HNC Cancer Quality of Q32 Question}
  \item{HN_Q33}{HNC Cancer Quality of Q33 Question}
  \item{HN_Q34}{HNC Cancer Quality of Q34 Question}
  \item{HN_Q35}{HNC Cancer Quality of Q35 Question}
  \item{HN_Q36}{HNC Cancer Quality of Q36 Question}
  \item{HN_Q37}{HNC Cancer Quality of Q37 Question}
  \item{HN_Q38}{HNC Cancer Quality of Q38 Question}
  \item{HN_Q39}{HNC Cancer Quality of Q39 Question}
  \item{HN_Q40}{HNC Cancer Quality of Q40 Question}
  \item{HN_Q41}{HNC Cancer Quality of Q41 Question}
  \item{HN_Q42}{HNC Cancer Quality of Q42 Question}
  \item{HN_Q43}{HNC Cancer Quality of Q43 Question}
  \item{HN_Q44}{HNC Cancer Quality of Q44 Question}
  \item{HN_Q45}{HNC Cancer Quality of Q45 Question}
  \item{HN_Q46}{HNC Cancer Quality of Q46 Question}
  \item{HN_Q47}{HNC Cancer Quality of Q47 Question}
  \item{HN_Q48}{HNC Cancer Quality of Q48 Question}
  \item{HN_Q49}{HNC Cancer Quality of Q49 Question}
  \item{HN_Q50}{HNC Cancer Quality of Q50 Question}
  \item{HN_Q51}{HNC Cancer Quality of Q51 Question}
  \item{HN_Q52}{HNC Cancer Quality of Q52 Question}
  \item{HN_Q53}{HNC Cancer Quality of Q53 Question}
  \item{HN_Q54}{HNC Cancer Quality of Q54 Question}
  \item{HN_Q55}{HNC Cancer Quality of Q55 Question}
  \item{HN_Q56}{HNC Cancer Quality of Q56 Question}
  \item{HN_Q57}{HNC Cancer Quality of Q57 Question}
  \item{HN_Q58}{HNC Cancer Quality of Q58 Question}
  \item{HN_Q59}{HNC Cancer Quality of Q59 Question}
  \item{HN_Q60}{HNC Cancer Quality of Q60 Question}
  \item{HN_Q61}{HNC Cancer Quality of Q61 Question}
  \item{HN_Q62}{HNC Cancer Quality of Q62 Question}
  \item{HN_Q63}{HNC Cancer Quality of Q63 Question}
  \item{HN_Q64}{HNC Cancer Quality of Q64 Question}
  \item{HN_Q65}{HNC Cancer Quality of Q65 Question}
  }
  #' @source <https://github.com/apstat/QoLMiss-Package>
}
\usage{
hnc_df
}
\description{
A simulated data for Head and Neck cancer Quality of Life.
}
\keyword{datasets}
