% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thyc_df_miss.R
\docType{data}
\name{thyc_df_miss}
\alias{thyc_df_miss}
\title{Thyroid cancer Quality of Life data with missing values.}
\format{
A data frame with 60 rows and 2 variables:
\describe{
  \item{ID}{Participant's identification}
  \item{time}{Time Variable}
  \item{event}{status as Variable}
  \item{arm}{Therapeutic Arm}
  \item{THY_Q31}{Thyroid Cancer Quality of Q31 Question}
  \item{THY_Q32}{Thyroid Cancer Quality of Q32 Question}
  \item{THY_Q33}{Thyroid Cancer Quality of Q33 Question}
  \item{THY_Q34}{Thyroid Cancer Quality of Q34 Question}
  \item{THY_Q35}{Thyroid Cancer Quality of Q35 Question}
  \item{THY_Q36}{Thyroid Cancer Quality of Q36 Question}
  \item{THY_Q37}{Thyroid Cancer Quality of Q37 Question}
  \item{THY_Q38}{Thyroid Cancer Quality of Q38 Question}
  \item{THY_Q39}{Thyroid Cancer Quality of Q39 Question}
  \item{THY_Q40}{Thyroid Cancer Quality of Q40 Question}
  \item{THY_Q41}{Thyroid Cancer Quality of Q41 Question}
  \item{THY_Q42}{Thyroid Cancer Quality of Q42 Question}
  \item{THY_Q43}{Thyroid Cancer Quality of Q43 Question}
  \item{THY_Q44}{Thyroid Cancer Quality of Q44 Question}
  \item{THY_Q45}{Thyroid Cancer Quality of Q45 Question}
  \item{THY_Q46}{Thyroid Cancer Quality of Q46 Question}
  \item{THY_Q47}{Thyroid Cancer Quality of Q47 Question}
  \item{THY_Q48}{Thyroid Cancer Quality of Q48 Question}
  \item{THY_Q49}{Thyroid Cancer Quality of Q49 Question}
  \item{THY_Q50}{Thyroid Cancer Quality of Q50 Question}
  \item{THY_Q51}{Thyroid Cancer Quality of Q51 Question}
  \item{THY_Q52}{Thyroid Cancer Quality of Q52 Question}
  \item{THY_Q53}{Thyroid Cancer Quality of Q53 Question}
  \item{THY_Q54}{Thyroid Cancer Quality of Q54 Question}
  \item{THY_Q55}{Thyroid Cancer Quality of Q55 Question}
  \item{THY_Q56}{Thyroid Cancer Quality of Q56 Question}
  \item{THY_Q57}{Thyroid Cancer Quality of Q57 Question}
  \item{THY_Q58}{Thyroid Cancer Quality of Q58 Question}
  \item{THY_Q59}{Thyroid Cancer Quality of Q59 Question}
  \item{THY_Q60}{Thyroid Cancer Quality of Q60 Question}
  \item{THY_Q61}{Thyroid Cancer Quality of Q61 Question}
  \item{THY_Q62}{Thyroid Cancer Quality of Q62 Question}
  \item{THY_Q63}{Thyroid Cancer Quality of Q63 Question}
  \item{THY_Q64}{Thyroid Cancer Quality of Q64 Question}
  }
  @source <https://github.com/apstat/QoLMiss-Package>
}
\usage{
thyc_df_miss
}
\description{
A simulated data for Thyroid cancer Quality of Life.
}
\keyword{datasets}
